refactor: move source code in src/
