+++ src/functions/updateAvatar.js
