+++ start.sh
@@ -2,7 +2,7 @@
 cd "$(dirname "$0")" || return
 
 while true; do
-	node .
+	npm start
 	echo Restarting in 10 seconds...
 	sleep 10
 done
